modularized the code, implemented three functions:
  * one_less_than(): for processing relation joins that are exactly one less than in nature
  * somewhat_less_than(): for processsing relation joins that are somewhat less than
  * eliminate_redundant_fragments(): clear out fragments that are redundant

Instead of a fixed repeat looping to ensure full propagation, I now take the MD5sum before and after
a run of the core loop, and compare. If NEWMD5 and OLDMD5 match, there was no change made, we can
exit the loop.

Still need to fully propagate all iterations to all valid possibilities. But: usefully close.